+++ apps/examples/async/main.c
@@ -307,12 +307,7 @@ static void input_task(void *param) {
 }
 
 int main(void) {
-    struct am_ao_state_cfg cfg = {
-        .on_idle = am_pal_on_idle,
-        .crit_enter = am_pal_crit_enter,
-        .crit_exit = am_pal_crit_exit
-    };
-    am_ao_state_ctor(&cfg);
+    am_ao_state_ctor(/*cfg=*/NULL);
 
     am_event_add_pool(
         m_event_pool,