+++ apps/examples/ringbuf/main.c
@@ -67,12 +67,7 @@ static void test_ringbuf_threading(void) {
 
     am_ringbuf_ctor(&g_ringbuf, buf, AM_COUNTOF(buf));
 
-    struct am_ao_state_cfg cfg_ao = {
-        .on_idle = am_pal_on_idle,
-        .crit_enter = am_pal_crit_enter,
-        .crit_exit = am_pal_crit_exit
-    };
-    am_ao_state_ctor(&cfg_ao);
+    am_ao_state_ctor(/*cfg=*/NULL);
 
     ringbuf_reader_ctor();
     ringbuf_writer_ctor();