+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [0-based versioning](https://0ver.org/).
 - Fix AM_GET_MACRO_2_() documentation.
 - Relocate compiler independent macros to `common/macros.h`.
 - Allow configuration set to NULL in `am_event_state_ctor()`.
+- Allow configuration set to NULL in `am_ao_state_ctor()`.
 
 ### Added